+++ releasenotes.md
@@ -9,6 +9,7 @@
 - `constdef` inference through binary expressions: `Foo f = Foo.AUDIO | Foo.VIDEO` can be written `Foo f = AUDIO | VIDEO;`
 - Fix for LLVM 22+ compatibility #2987
 - `@weaklink` for just affecting linkage.
+- Add a fully static build of `c3c` for Linux. #2949
 
 ### Stdlib changes
 - Add contract on `any_to_enum_ordinal` and `any_to_int` to improve error when passed an empty any. #2977